ai-service/requirements.txt
# OCR and Text Extraction
pytesseract
PyMuPDF
python-docx
Pillow

# AI/ML and Data Processing
pandas
numpy
scikit-learn
spacy

# Service, LLMs, and utilities
fastapi
uvicorn
google-generativeai
httpx
python-dotenv
joblib
